requirements.txt
fastapi
uvicorn
pandas
scikit-learn
mysql-connector-python
joblib
